% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_functions.R
\name{create_card}
\alias{create_card}
\title{Create html card using R}
\usage{
create_card(person_page, img_src, headline, alt, name, color1)
}
\description{
This is the underlying function, written by Dax
}
