% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card_functions.R
\name{create_card_from_qmd}
\alias{create_card_from_qmd}
\title{Read yaml data and use to create card}
\usage{
create_card_from_qmd(file)
}
\arguments{
\item{file}{path to an 'article' qmd file}
}
\description{
This is in early dev
}
