% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_extract.R
\name{name_extract}
\alias{name_extract}
\title{name_extract}
\usage{
name_extract(
  path,
  verify = FALSE,
  language = NULL,
  no_bayes = FALSE,
  odds_details = FALSE,
  data_sources = NULL,
  tokens = NULL
)
}
\arguments{
\item{path}{(character) path to a file}

\item{verify}{(logical) verify names? default: `FALSE`}

\item{language}{(logical) text's language. by default it's
automatically detected. default: `NULL`}

\item{no_bayes}{(logical) do not run Bayes algorithms. default: `FALSE`}

\item{odds_details}{(logical) show details of odds calculation.
default: `FALSE`}

\item{data_sources}{(numeric vector) IDs of data sources to display for
matches. default: `NULL`. e.g., `c(1, 11, 179)`}

\item{tokens}{(logical) xxx. default: `NULL`}
}
\value{
list of two data.frames:

- meta: metadata
- names: names and their parts, varies based on function parameters
}
\description{
extract names using gnfinder. Note that this is taken (near-)verbatim from 
the `namext` package by Scott Chamberlain, accessed at 
https://github.com/ropensci-archive/namext on 2023-02-15.
}
