% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_functions.R
\name{taxon_frequencies}
\alias{taxon_frequencies}
\title{functions to support taxonomic data extraction from PDFs}
\usage{
taxon_frequencies(file_path)
}
\description{
This is a test for building documentation
}
