% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_attachments}
\alias{zotero_attachments}
\title{Zotero attachments listed under a given itemKey}
\usage{
zotero_attachments(user_id, group_id, api_key, item_key)
}
\description{
Zotero attachments listed under a given itemKey
}
