% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero_api.R
\name{zotero_items_by_tag}
\alias{zotero_items_by_tag}
\title{Zotero items listed under a specific tag}
\usage{
zotero_items_by_tag(user_id, group_id, api_key, tag)
}
\description{
Zotero items listed under a specific tag
}
